backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
psycopg[binary]==3.2.9
PyJWT==2.10.1
prometheus-client==0.22.1
httpx==0.28.1
websockets==15.0.1
